% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{baseline}
\alias{baseline}
\title{Baseline}
\usage{
baseline(x, dates, nyear = Inf, center = FALSE,
  smoothing.fun = c("ma", "spline", "none"), ...)
}
\arguments{
\item{x}{Numeric vector. Contains data from which the baseline is computed.}

\item{dates}{POSIXt vector the same length as \code{x}. Dates corresponding
to each element of \code{x}. Alternatively, a character vector can be
provided as long as it can be coerced to a POSIXt object.}

\item{nyear}{Integer. Number of years around the current one to consider 
in the calendar mean step.}

\item{center}{Logical. Is the calendar mean of step 2 centered on current 
year? If FALSE (the default), the past \code{nyear} are used.}

\item{smoothing.fun}{One of \code{"ma"}, \code{"spline"} or \code{"none"}.
Type of smoothing to perform on step 1. If \code{"ma"} (the default), 
a moving average is performed using the function \code{\link[forecast]{ma}}
from the package \code{forecast}. If \code{"spline"}, 
\code{\link[stats]{smooth.spline}} is used. If \code{"none"}, 
no smoothing is performed.}

\item{...}{Additional arguments to be passed to function chosen in 
\code{smoothing.fun}.}
}
\value{
A numeric vector the same length as \code{x}.
}
\description{
Compute a baseline from a series through two steps: 1) smoothing and 
  2) calendar mean.
}
\details{
Computes a baseline of expected mortality (or other health issue) 
  to use as a reference to later compute a series of excess mortality. 
  The baseline is computed through two steps:
  \enumerate{
    \item Smoothing of the series \code{x} ;
    \item Computing of a calendar mean, \emph{i.e.} each day of year of 
       the baseline is the mean 
       of the smoothed series at the same day of year for \code{nyear}.
  }
}
\examples{
  data(chicagoNMMAPS)
  x <- chicagoNMMAPS$death
  dates <- as.POSIXlt(chicagoNMMAPS$date)

  em <- baseline(x, dates, order = 15)
  plot(dates, x)
  lines(dates, em, col = "red")

  em2 <- baseline(x, dates, smoothing.fun = "spline")
  plot(dates, x)
  lines(dates, em2, col = "red")

  em3 <- baseline(x, dates, nyear = 2)
  plot(dates, x)
  lines(dates, em3, col = "red")
}
\references{
Chebana F., Martel B., Gosselin P., Giroux J.X., Ouarda T.B.M.J., 2013. 
    A general and flexible methodology to define thresholds for heat health 
    watch and warning systems, applied to the province of Quebec (Canada). 
    International journal of biometeorology 57, 631-644.
}
\seealso{
[excess()] to compute excesses from the baseline.
}
