% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excess.R
\name{excess}
\alias{excess}
\alias{excess_inverse}
\title{Excess from a baseline}
\usage{
excess(x, expected, fact = 100, ...)

excess_inverse(excess, expected, fact)
}
\arguments{
\item{x}{Numeric vector. Contains original data.}

\item{expected}{Numeric vector the same length as \code{x}. The baseline 
used in the computation of excesses. If missing in \code{excess}, the  
function \code{\link{baseline}} is called with the additional arguments 
passed in \code{...}. If missing in \code{excess_inverse}, the attribute 
\code{baseline} is used if present.}

\item{fact}{Numeric. A factor to apply to the relative excess. Default to 100
to express excesses in percentage.}

\item{...}{Arguments to be passed to \code{\link{baseline}} if 
\code{expected} is missing. See \code{?baseline}.}

\item{excess}{Numeric vector. A series of excesses to be reverted to the 
original scale of \code{x}.}
}
\value{
\code{excess}: A numeric vector of class \code{excess} the same 
  length as \code{x}.In addition, contains the attributes \code{baseline} and
  \code{fact} to keep track of the computation.

\code{excess_inverse}: A numeric vector the same length as 
  \code{excess}.
}
\description{
\code{excess} computes the relative excesses of \code{x} relatively to a 
  baseline. \code{excess_inverse} reverts the process.
}
\details{
Compute relative excesses from a baseline. Excesses are computed as:
  \deqn{excess = (x - expected) / expected * fact}
}
\examples{
  library(dlnm)
  data(chicagoNMMAPS)
  x <- chicagoNMMAPS$death
  dates <- as.POSIXlt(chicagoNMMAPS$date)

  om <- excess(x, dates = dates, order = 15)
  xrec <- excess_inverse(om)

  plot(dates, x)
  lines(dates, xrec, col = "red")

}
\references{
Chebana F., Martel B., Gosselin P., Giroux J.X., Ouarda T.B.M.J., 2013. 
    A general and flexible methodology to define thresholds for heat health 
    watch and warning systems, applied to the province of Quebec (Canada). 
    International journal of biometeorology 57, 631-644.
}
\seealso{
\code{\link{baseline}} for computing expected values.
}
