% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.threshold.R
\name{find.threshold}
\alias{find.threshold}
\title{Find thresholds}
\usage{
find.threshold(indicators, episodes, u.grid, fixed.alphas = NULL,
  alpha.step = 0.1, decreasing.alphas = TRUE, same.alphas = TRUE,
  order.result = NULL, order.decreasing = TRUE, r = NULL,
  trim = NULL, thinning = c("none", "days", "episodes"),
  progressBar = FALSE)
}
\arguments{
\item{indicators}{List of matrices containing the lagged covariates to 
construct weighted indicators. Can also be a matrix in the case of a
single indicator.}

\item{episodes}{A matrix or data.frame containing the indices of extreme
episodes. Must contain the time indices in the first column and the 
corresponding episode index in the second column as returned by  
\code{\link{episodes}}.}

\item{u.grid}{A list of vectors containing the grid of thresholds to be 
considered. The list must have the same number of elements than 
\code{indicators}. If missing, the default is to use all percentiles
from 0.8 to 1.}

\item{fixed.alphas}{A list of optional prefixed weightings for a subset of
indicators. When provided, must have the same length as \code{indicators}
and contain \code{NULL} for each non-fixed indicator.}

\item{alpha.step}{Numeric value between 0 and .5. The step of the
sequence of weightings tested for the indicator.}

\item{decreasing.alphas}{Logical. If TRUE (the default), the alpha 
weightings are constrained to decrease with the lag.}

\item{same.alphas}{Logical. If TRUE, the weightings are constrained to be
identical for each indicator. Note that trying different weightings
(\code{same.alpha = FALSE}) can lead to significantly longer
computation time.}

\item{order.result}{Character or numeric value indicating a column used
to order the returned table. Can also be a vector.}

\item{order.decreasing}{Logical. If TRUE (the default), ordering of the
table is made by decreasing order to the column specified by
\code{order.result}.}

\item{r}{Positive integer. Number of consecutive values below threshold
following an excess to end the episode. By default, take the attribute
\code{r} in \code{episodes} or 1 if absent.}

\item{trim}{Positive integer. If not \code{NULL} (the default), the first
\code{trim} lines of the result are returned by the function. Depends on 
the parameter \code{order.result}.}

\item{thinning}{Character string indicating if the results should be thinned
before returning. When \code{thinning = "none"} (the default), all tested
combinations are returned. When \code{thinning = "days"}, combinations 
for which there is at least another combinations such that specificity
and sensivity are both higher, are removed from the returned results.
When \code{thinning = "episodes"} the thinning is performed on the basis
of episodes instead of days.}

\item{progressBar}{Logical indicating if a progressBar is displayed during
execution of the function. If TRUE, may greatly increase execution time.}
}
\value{
A data.frame containing a subset (unless \code{thinning = "none"})  
   of weightings and thresholds. Weightings correspond to the columns
   with a name containing "alpha" and threshold to names beginning with 
   "threshold". In addition, several scores are given in each line:
     \item{Detected}{The number of indices in \code{episodes} detected by
       the combination.}
     \item{Missed}{The number of indices in \code{episodes} missed by
       the combination.}
     \item{Sensitivity}{The proportion of indices in \code{episodes} detected by
       the combination, \emph{i.e.} Detected / n.}
     \item{False_alarms}{The number of false alarms, \emph{i.e.} of indices
       found by the combination which are not in \code{episodes}.}
     \item{Specificity}{The proportion of false alarms, \emph{i.e.} 
       False_alarms / n.}
     \item{Episodes_found}{The number of episodes found. An episode is
       found when at least one of its days is found.}
     \item{Episodes_sensitivity}{The proportion of episodes found.}
     \item{False_episodes}{The number of false episodes found, i.e. absent
       from the provides \code{episodes}.}
}
\description{
Evaluate different combinations of threshold and indicator weightings
   according to their sensitivity and specificity.
}
\details{
We consider a warning system as a couple indicator/threshold
   used to launch alerts when forecasts of the indicator exceed the
   threshold. In the present function, the indicators considered are
   linear combinations of all matrix columns in the parameter 
   \code{indicators}, with the constraint that, for each indicator, the
   weights sum to 1.
   
   The indicator and threshold are determined by evaluating a large range
   of different weightings and threshold (given in \code{u.grid} and 
   \code{alpha.step}). For each combination of indicators/thresholds,
   the function computes the indices corresponding to alerts in the data
   and compare them to the actual values given in \code{episodes}. If
   \code{thinning != "none"}) the function then selects a subset of best 
   candidates for which the number of detected days are maximum while the
   number of false alarms is minimum. Note that if 
   \code{thinning = "episodes"}, the subset is selected on the basis of
   detected and false episodes instead of days. It is left to the user to 
   choose the best combination by a trade-off between specificity and 
   sensitivity.
}
\examples{
   library(dlnm)
   data(chicagoNMMAPS)
   x <- chicagoNMMAPS$death
   dates <- as.POSIXlt(chicagoNMMAPS$date)
   n <- nrow(chicagoNMMAPS)

   # Compute over-mortality
   om <- excess(x, dates = dates, order = 15)
  
   # Extract all days for which om is above 40\%
   epis <- episodes(om, u = 40)

   # Prepare indicator based on temperature until lag 2
   indic <- matrix(NA, nrow = n, ncol = 3)
   indic[,1] <- chicagoNMMAPS$temp  # lag 0
   indic[,2] <- c(NA, chicagoNMMAPS$temp[-n]) # Lag 1
   indic[,3] <- c(NA, NA, chicagoNMMAPS$temp[1:(n-2)]) # lag 2

   # Evaluate different threshold/indicators based on these episodes
   find.threshold(indic, epis, u.grid = 20:35, thinning = "episodes",
     order.result = "Episodes_found")

}
\references{
Chebana F., Martel B., Gosselin P., Giroux J.X., Ouarda T.B.M.J., 2013. 
     A general and flexible methodology to define thresholds for heat health 
     watch and warning systems, applied to the province of Quebec (Canada). 
     International journal of biometeorology 57, 631-644.

   Pascal M., Laaidi K., Ledrans M., Baffert E., Caserio-Schonemann C.,
     Le Tertre A., Manach J., Medina S., Rudant J., Empereur-Bissonnet P.,
     2006. France's heat health watch warning system. International journal
     of biometeorology 50, 144-153.
}
\seealso{
\code{\link{episodes}} for extracting episodes of extreme values and 
   \code{\link{predict_alarms}} for alarms prediction.
}
